+++ packages/atproto_core/lib/src/client/anonymous_client.dart
@@ -19,7 +19,7 @@ class AnonymousClient implements Client {
     required final String service,
     final Map<String, dynamic>? parameters,
     required final xrpc.To<T> to,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
     required final Duration timeout,
     final xrpc.GetClient? getClient,
   }) async =>
@@ -29,7 +29,7 @@ class AnonymousClient implements Client {
         service: service,
         parameters: parameters,
         to: to,
-        converter: converter,
+        adaptor: adaptor,
         timeout: timeout,
         getClient: getClient,
       );
@@ -84,15 +84,13 @@ class AnonymousClient implements Client {
     final String? service,
     final Map<String, dynamic>? parameters,
     final xrpc.To<T>? to,
-    final xrpc.Decoder? decoder,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   }) =>
       xrpc.subscribe(
         methodId,
         service: service,
         parameters: parameters,
         to: to,
-        decoder: decoder,
-        converter: converter,
+        adaptor: adaptor,
       );
 }