+++ packages/atproto_core/lib/src/client/auth_required_client.dart
@@ -22,7 +22,7 @@ class AuthRequiredClient implements Client {
     required final String service,
     final Map<String, dynamic>? parameters,
     required final xrpc.To<T> to,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
     required final Duration timeout,
     final xrpc.GetClient? getClient,
   }) async =>
@@ -33,7 +33,7 @@ class AuthRequiredClient implements Client {
         parameters: parameters,
         headers: {'Authorization': 'Bearer $_accessJwt'},
         to: to,
-        converter: converter,
+        adaptor: adaptor,
         timeout: timeout,
         getClient: getClient,
       );
@@ -94,8 +94,7 @@ class AuthRequiredClient implements Client {
     final String? service,
     final Map<String, dynamic>? parameters,
     final xrpc.To<T>? to,
-    final xrpc.Decoder? decoder,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   }) =>
       throw UnimplementedError();
 }