+++ packages/xrpc/lib/xrpc.dart
@@ -17,7 +17,7 @@ export 'package:xrpc/src/protocol.dart';
 export 'package:xrpc/src/serializable.dart';
 export 'package:xrpc/src/subscription.dart';
 export 'package:xrpc/src/xrpc.dart'
-    show query, procedure, upload, subscribe, To, Decoder, JsonConverter;
+    show query, procedure, upload, subscribe, To, ResponseAdaptor;
 export 'package:xrpc/src/xrpc_error.dart';
 export 'package:xrpc/src/xrpc_request.dart';
 export 'package:xrpc/src/xrpc_response.dart';